tox: disable W504 "Line break occurred after a binary operator"
This is a new coding rule, which actively contradicts the previous
pep8 rule, W503 "line break before binary operator".  Both rules
are currently in force, which doesn't make a lot of sense.

Until the standards people work this out, I'm arbitrarily choosing
to disable the rule that involes the fewest coding changes.